requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.10.1
cryptography==44.0.0
